+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add Google Calendar Event invitations from Google Mail as a notification
 - Support multiple authentication mechanisms (ie. local + Google)
 - Support Passkey authentication
+- Add notifications pagination, filtering and sorting
 
 ### Changed